Removed exception handling from handler and minor fixes

+++ LIMS.Application/Handlers/Technician/TechnicianCommandHandler/DeleteTechnicianCommandHandler.cs
@@ -19,21 +19,19 @@ namespace LIMS.Application.Handlers.Technician.TechnicianCommandHandler
 
         public async Task<Unit> Handle(DeleteTechnicianCommand request, CancellationToken cancellationToken)
         {
-            try
-            {
-                var entityToDelete = await _technician
                 var technicianEntity = AutoMapperConfiguration.Mapper.Map<Domain.Models.Technician>(request.TechnicianId);
                 if (technicianEntity == null)
                 {
-                    throw new Exception
+                    throw new ApplicationException("Unable to map due to an issue with mapper.");
                 }
-                var response = await _technicianCommandRepository.DeleteAsync(technicianEntity.TechnicianId);
+
+                var entityToDelete = await _technicianQueryRepository.GetAsyncById(technicianEntity.TechnicianId, Domain.Common.DataTables.TechnicianTable, Domain.Common.DataColumns.TechnicianId);
+                if(entityToDelete != null)
+                {
+                   await _technicianCommandRepository.DeleteAsync(technicianEntity.TechnicianId);
+                }
                 return response;
-            }
-            catch (Exception ex)
-            {
-                throw new 
-            }
+            
 
         }
     }
